fix: ts column type
